@@ -1,9 +1,12 @@
 #!/usr/bin/env node
 
-const fs = require('fs')
-const path = require('path')
+import fs from 'fs'
+import path from 'path'
 
-const pkgUp = require('pkg-up')
+import pkgUp from 'pkg-up'
+
+import { fileURLToPath } from 'url'
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
 const TSCONFIG_JSON_CONTENT = `{
   "extends": "@chatie/tsconfig",
